+++ src/main/java/com/almightyalpaca/jetbrains/plugins/discord/settings/DiscordIntegrationSettingsPanel.java
@@ -274,7 +274,7 @@ public class DiscordIntegrationSettingsPanel
 
     private Path createFolder(@NotNull Path path)
     {
-        if (applicationDebugLoggingEnabled.isEnabled() && verifyLogFolder() && !Files.isDirectory(path))
+        if (this.applicationDebugLoggingEnabled.isSelected() && verifyLogFolder() && !Files.isDirectory(path))
         {
             try
             {