+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## development
+- Updated Java-DiscordRPC to v2.0.0
 - Added Slim and CoffeeScript (#34)
 - Fixed file not reappearing after unsetting project description (#31)
 - Fixed debug folder being created even when debug logging was disabled